refactor: separate app setup and server for testing
